Adding update-changelog command
