+++ buildSrc/src/main/kotlin/io/getstream/chat/android/command/changelog/plugin/ChangelogCommandExtension.kt
@@ -1,0 +1,6 @@
+package io.getstream.chat.android.command.changelog.plugin
+
+open class ChangelogCommandExtension {
+    var changelogPath: String = "CHANGELOG.md"
+    var changelogModel: String = "CHANGELOG_MODEL.md"
+}
