+++ buildSrc/src/main/kotlin/io/getstream/chat/android/command/release/plugin/ReleaseCommandExtension.kt
@@ -2,5 +2,4 @@ package io.getstream.chat.android.command.release.plugin
 
 open class ReleaseCommandExtension {
     var changelogPath: String = "CHANGELOG.md"
-    var changelogModel: String = "CHANGELOG_MODEL.md"
 }
